Ownership lives with its validator; boot rejects aliased entries
The package map is the loader-validated install claim, so it moves into the
loader — registry.py returns to the capability catalogs it always was, and
workflows.py asks the loader who owns a module. iter_extensions() gains the
entry-key/name check _resolve() already had: production boot no longer accepts
an aliased entry the app-less path rejects.
