+++ backend/tests/test_extension_doctor_checks.py
@@ -31,7 +31,7 @@ def external_package():
     load mutates so the rest of the suite sees the in-tree extensions untouched.
     Mirrors the proof-extension suite's fixture."""
     from blinker import signal
-    from druks.extensions import registry as extensions_registry
+    from druks.extensions import loader as extensions_loader
     from druks.extensions.registry import agents, webhooks, workflows
     from druks.models import Base
 
@@ -39,7 +39,7 @@ def external_package():
 
     tables = set(Base.metadata.tables)
     registries = {registry: dict(registry._items) for registry in (agents, webhooks, workflows)}
-    packages = dict(extensions_registry._workflow_packages)
+    packages = dict(extensions_loader._workflow_packages)
     finished = signal("run.finished")
     receivers = dict(finished.receivers)
     try:
@@ -50,8 +50,8 @@ def external_package():
             Base.metadata.remove(Base.metadata.tables[name])
         for registry, snapshot in registries.items():
             registry._items = snapshot
-        extensions_registry._workflow_packages.clear()
-        extensions_registry._workflow_packages.update(packages)
+        extensions_loader._workflow_packages.clear()
+        extensions_loader._workflow_packages.update(packages)
         finished.receivers = receivers
         for name in [m for m in sys.modules if m == _PACKAGE or m.startswith(f"{_PACKAGE}.")]:
             del sys.modules[name]
